Be 7 add validation (#7)
* deleting your own account has been implemented

* Restrict user update and delete actions to authorized users

Added checks to ensure that users can only update or delete their own profiles.

* upd ci/cd

---------

+++ poll/db/model_users.py
@@ -90,6 +90,7 @@ class UserRepository:
         logger.info("Deleting user: %s", user_id)
         result = await self.session.execute(select(User).filter(User.id == user_id))
         user = result.scalars().first()
-        await self.session.delete(user)
-        await self.session.commit()
+        if user:
+            await self.session.delete(user)
+            await self.session.commit()
         return user
